Fixed Menu Selection Logging
1. Menu Selection Logging now correctly logs each object in the menu
selection with the associated interaction type.
TODOS: Working on adding a better way to distinguish the relationship
between the menu items currently if a menu item has the same objects
and interaction type there is no way to distinguish them. Idea is to
extract the action from the relationship and log that.

Also enable the displayMenu computer action, currently only selection
is enabled which is the user action of selecting a menu. Displaymenu
will be the computer action upon correct user action of moving an
object to the correct other object location.