Changing terminology, expanding menu display and selection details, simplifying log output data
1. Changed ContextualMenuDataSource and MenuItemDataSource to now also
record the relationship associated with that menu item. To do so I also
had to change the function simulatePossibleInteractionForMenuItem in
pmviewcontroller and possibleinteractions. Possbieinteractions now
creates the array that holds all relationships for the possible
interactions for the correct possible solutions.
2. Added variables to hold collision objects and most recent
relationships parsed for easier access for logging which objects are
being collided into.
3. Changed the structure of the display menu and selection of menu item
functions to now record relationships to better distinguish which menu
items are which for menu items that have the same interaction and
images.
Todo:
1. More code cleanup (mostly removing slog debugging statements)
2. Further Simplification of logging structure format and making sure
input, selection and action fields are intuitive and consistent
3. Testing emergency swipe logging
4. Testing Word and object selection logging in introduction
5. Testing Multiple pages logging
6. Testing audio logging for language conditions
7. Complete test of all chapters