Added group auto after appear functionality, framework for checkleft, checkright, checktop, checkdown, added black sentence override
1. Added a height parameter to alternative images.
If an alternative image does not have a height specified it will call
the correct javascript function.
ImageManipulation.js now has an overloaded function for swapping and
appearing images.

2. Added solution steps for checkLeft, checkRight, checkDown, checkTop
Added a new variable that stores an objects ending location after it is
calculated using the restraints of the screen. This is then used to
compare and objects starting location and ending location to confirm
the check solution step.

3. Added an override to make what would normally be blue sentences
black. simply add “black” to “sentence actionSentence black” to keep
the sentence black.