he Navajo Hogan, second pm first sentence "the frame of a hogan is made form tree branches" is being duplicated and causing the ideas/solutions to be off by 1 and causing the app to crash (Jithin) FIXED - I didnt see that issue, but saw an error in meta data file maybe that caused it.
Looks like there is a mapping issue for The Inuit Igloo on the sentence "to build an igloo, the inuit people use a shovel to make a circle in the snow", user is able to press next without manipulation (Jithin) FIXED

There is another issue on The Sioux Teepee, for the sentence "the sioux hunted buffalo with bows and arrows( tap the arrow)" causes a system crash when attempting the manipulation (Jithin) FIXED

ITS004 Remove Spanish instuction audio at the beginning of Native Homes introduction FIXED

ITS004 The first manipulation in The Navajo Hogan is no longer correct. The mudpile should dissapear and the picture of the hogan should swap out with the hogan that is completed. FIXED

ITS004 Typo on last sentence in The Inuit Igloo. Extra period at the end of "There is no need for a fire!." FIXED

ITS004: Second time going through The Sioux Teepee, the sentence, "The Sioux put the wrapped up teepee onto a sled." was black and did not require a manipulation. But it worked the first time through. I made a couple syntax errors the first time through, but never received audio of the sentence. Oh, I see now that the following sentence, "A horse pulled the sled" is blue, but it should be black. Here, the app  wants me to move the Sioux to put the teepee on the sled. FIXED

ITS004: The last page of Sioux Teepee is missing text so cannot progress. FIXED